Player to member entity changes.  Add working for event results
