+++ src/DartLeague/DartLeague.Repositories/LeagueData/DartEventResult.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace DartLeague.Repositories.LeagueData
 {
@@ -11,5 +13,6 @@ namespace DartLeague.Repositories.LeagueData
         public int OrderId { get; set; }
         public int MemberId { get; set; }
         public string SpecificEventName { get; set; }
+        public Member Member { get; set; }
     }
 }
