+++ src/DartLeague/DartLeague.Repositories/LeagueData/LeagueContext.cs
@@ -135,12 +135,14 @@ namespace DartLeague.Repositories.LeagueData
 
                 entity.Property(e => e.MemberId)
                     .HasColumnName("memberId")
-                    .HasColumnType("int(11)");
+                    .HasColumnType("int(10) unsigned");
 
                 entity.Property(e => e.SpecificEventName)
                     .IsRequired()
                     .HasColumnName("specificEventName")
                     .HasColumnType("varchar(255)");
+
+                entity.HasOne(e => e.Member).WithMany();
             });
             modelBuilder.Entity<DartEvent>(entity =>
             {
