+++ src/DartLeague/DartLeague.Web/Areas/Site/Views/DartEvent/Result.cshtml
@@ -7,7 +7,7 @@
 <h2>Dart Event Results</h2>
 <div><b>@Model.DartEventName</b> - @Model.DartEventDate</div>
 <div class="table-responsive">
-    <form asp-controller="EventResult" asp-action="Index" method="post">
+    <form asp-controller="DartEvent" asp-action="Result" method="post">
         <table class="table table-striped table-hover">
             <thead>
                 <tr>
@@ -32,7 +32,7 @@
                             <td style="vertical-align: middle">@result.SpecificEventName</td>
                             <td style="vertical-align: middle">@result.MemberName</td>
                             <td style="vertical-align: middle">@result.Finished</td>
-                            <td><a class="btn btn-link confirmation" asp-area="Site" asp-controller="EventResult" asp-action="Delete" asp-route-id="@result.Id" title="Delete"><i class="fa fa-trash-o fa-lg"></i></a>
+                            <td><a class="btn btn-link confirmation" asp-area="Site" asp-controller="DartEvent" asp-action="DeleteResult" asp-route-id="@result.Id" title="Delete"><i class="fa fa-trash-o fa-lg"></i></a>
                         </tr>
                     }
                 }
